Burn Down Charted Added
When you open it, go to sheet 2 of the workbook. There you enter in how
many "task hours" you've completed on the appropriate day (note how I
said "task hours" and not "hours", treat it more like a percentage of
how much of the task you've done...if that makes sense.) It should
automatically update the iteration a burndown chart below the table.